# docs/openapi-skeleton.yaml
openapi: 3.0.3
info:
  title: Loadline
  version: 0.1.0
  description: >-
    Premium-composition platform for life and health risk. Every production model
    declares the term of the price it acts on — expected claims, loading for risk,
    or loading for expense — every rating factor carries a per-jurisdiction
    permitted-use position, and no offer may leave the permitted variation
    envelope. Engagement touchpoints and realised cohort experience decide whether
    an expected-claims saving may be credited to price.
  x-product-slug: ai-ins-loadline
  x-source: ai-in-insurance/h2o_ai-session2-h2oaiworldlondon2018-madsenv2-181102025107/
servers:
  - url: https://api.example.com
...
paths:
...
  /v1/variation-envelopes:
    get:
      operationId: listVariationEnvelopes
      tags: [Composition]
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - name: jurisdiction
          in: query
          schema:
            type: string
            example: NL
        - name: productCode
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Permitted magnitude and frequency of price variation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListEnvelopeVariationEnvelope"
    post:
      operationId: setVariationEnvelope
      tags: [Composition]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VariationEnvelopeCreate"
      responses:
        "201":
          description: Envelope recorded with its filing or rule basis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataEnvelopeVariationEnvelope"
  /v1/permitting/factor-permits:
    get:
      operationId: listFactorPermits
      tags: [Permitting]
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - name: jurisdiction
          in: query
          schema:
            type: string
        - name: position
          in: query
          schema:
            type: string
            enum: [permitted, permitted_with_conditions, prohibited, undetermined]
        - name: dataCategory
          in: query
          schema:
            $ref: "#/components/schemas/DataCategory"
      responses:
        "200":
          description: Per-factor, per-jurisdiction permitted-use positions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListEnvelopeFactorPermit"
    post:
      operationId: recordFactorPermit
      tags: [Permitting]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FactorPermitCreate"
      responses:
        "201":
          description: Position recorded by a named compliance officer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataEnvelopeFactorPermit"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
...
  /v1/experience/cohorts:
    get:
      operationId: listCohortDefinitions
      tags: [Experience]
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - name: programmeId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Engaged and matched unengaged cohort definitions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListEnvelopeCohortDefinition"
    post:
      operationId: defineCohort
      tags: [Experience]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CohortDefinitionCreate"
      responses:
        "201":
          description: Cohort defined with its selection-effect separation method
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataEnvelopeCohortDefinition"
        "422":
          description: >-
            Rejected when an engaged cohort has no comparable unengaged cohort or
            no stated method for separating self-selection from behaviour.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
...
  responses:
...
    UnprocessableEntity:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
    ResponseMeta:
      type: object
      required: [requestId, timestamp]
      properties:
        requestId:
          type: string
        timestamp:
          type: string
          format: date-time
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
...
    DataCategory:
      type: string
      enum:
        - connected_life
        - connected_car
        - connected_home
        - connected_finance
        - medical_evidence
        - genetic
        - epigenetic
        - microbiome
        - geographic
        - declared_lifestyle
        - population_table
...
    VariationEnvelope:
      type: object
      required: [id, jurisdiction, productCode]
      properties:
        id:
          type: string
        jurisdiction:
          type: string
        productCode:
          type: string
        maxIncreasePercent:
          type: number
          format: float
        maxDecreasePercent:
          type: number
          format: float
        minRevisionIntervalDays:
          type: integer
          description: Shortest interval at which the filed structure permits change
        prohibitedDrivers:
          type: array
          items:
            type: string
          example: [inferred_price_sensitivity, prior_shopping_behaviour]
        basis:
          type: string
          enum: [rate_filing, regulatory_rule, policy_condition, treaty_term]
        basisReference:
          type: string
        effectiveFrom:
          type: string
          format: date
    VariationEnvelopeCreate:
      type: object
      required: [jurisdiction, productCode, basis]
      properties:
        jurisdiction:
          type: string
        productCode:
          type: string
        maxIncreasePercent:
          type: number
          format: float
        maxDecreasePercent:
          type: number
          format: float
        minRevisionIntervalDays:
          type: integer
        prohibitedDrivers:
          type: array
          items:
            type: string
        basis:
          type: string
        basisReference:
          type: string
...
    FactorPermit:
      type: object
      required: [id, factor, jurisdiction, position]
      properties:
        id:
          type: string
        factor:
          type: string
          example: residential_postcode_mortality_index
        dataCategory:
          $ref: "#/components/schemas/DataCategory"
        jurisdiction:
          type: string
        position:
          type: string
          enum: [permitted, permitted_with_conditions, prohibited, undetermined]
        proxyRiskAssessment:
          type: string
          description: Whether the factor proxies a protected characteristic
        restrictedHealthDataBasis:
          type: string
          description: Statutory or moratorium position for genetic and similar data
        explainableToApplicant:
          type: boolean
        conditions:
          type: array
          items:
            type: string
        officerId:
          type: string
        reviewedAt:
          type: string
          format: date-time
        nextReviewDue:
          type: string
          format: date
    FactorPermitCreate:
      type: object
      required: [factor, jurisdiction, position, officerId]
      properties:
        factor:
          type: string
        dataCategory:
          $ref: "#/components/schemas/DataCategory"
        jurisdiction:
          type: string
        position:
          type: string
          enum: [permitted, permitted_with_conditions, prohibited, undetermined]
        proxyRiskAssessment:
          type: string
        restrictedHealthDataBasis:
          type: string
        explainableToApplicant:
          type: boolean
        conditions:
          type: array
          items:
            type: string
        officerId:
          type: string
...
    CohortDefinition:
      type: object
      required: [id, programmeId, cohortKind]
      properties:
        id:
          type: string
        programmeId:
          type: string
        cohortKind:
          type: string
          enum: [engaged, matched_unengaged]
        matchedOn:
          type: array
          items:
            type: string
          example: [attained_age, sex, product, underwriting_class, duration]
        selectionSeparationMethod:
          type: string
          enum: [propensity_matching, stratified_comparison, randomised_offer, instrumental_variable]
        memberCount:
          type: integer
        observationWindowMonths:
          type: integer
        comparableCohortId:
          type: string
    CohortDefinitionCreate:
      type: object
      required: [programmeId, cohortKind, selectionSeparationMethod]
      properties:
        programmeId:
          type: string
        cohortKind:
          type: string
          enum: [engaged, matched_unengaged]
        matchedOn:
          type: array
          items:
            type: string
        selectionSeparationMethod:
          type: string
        observationWindowMonths:
          type: integer
        comparableCohortId:
          type: string
...
    DataEnvelopeVariationEnvelope:
      type: object
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/VariationEnvelope"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
    DataEnvelopeFactorPermit:
      type: object
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/FactorPermit"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
...
    DataEnvelopeCohortDefinition:
      type: object
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/CohortDefinition"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
...
    ListEnvelopeVariationEnvelope:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/VariationEnvelope"
            nextCursor:
              type: string
        meta:
          $ref: "#/components/schemas/ResponseMeta"
    ListEnvelopeFactorPermit:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/FactorPermit"
            nextCursor:
              type: string
        meta:
          $ref: "#/components/schemas/ResponseMeta"
...
    ListEnvelopeCohortDefinition:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/CohortDefinition"
            nextCursor:
              type: string
        meta:
          $ref: "#/components/schemas/ResponseMeta"
...