# platform/api-server/src/domains/composition/handlers/variation-envelope.handlers.ts
/**
 * Handlers for VariationEnvelope
 *
 * Generated by Handler Generator v2.0.0
 * Generator Version: 2.0.0
 * Domain: composition
 *
 * ⚠️  DO NOT EDIT THIS FILE MANUALLY
 * This file is auto-generated. Any manual changes will be overwritten.
 */

import type { FastifyRequest, FastifyReply } from "fastify";
import type { CompositionDomainModule } from "../dependencies/composition-ddd.dependencies.js";
import type { ListVariationEnvelopesInput, SetVariationEnvelopeInput } from "@loadline/services/composition";
// Use cases are available via deps.useCases.<repoKey>.<verb>
// Execution context is automatically set up by route wrapper (wrapHandlerInExecutionContext)

/**
 * Handler for GET /v1/variation-envelopes
 * Operation: listVariationEnvelopes
 */
export async function listVariationEnvelopes(
  request: FastifyRequest<{ Params: Record<string, string>; Querystring: { cursor?: string | string[], limit?: string | string[], jurisdiction?: string | string[], productCode?: string | string[] }; Body: unknown }>,
  reply: FastifyReply,
  deps: CompositionDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: ListVariationEnvelopesInput = {
    ...(request.query ?? {}),
  };
  const result = await deps.useCases.variationEnvelopes.list.execute(input);
  return reply.code(200).send(result);
}


/**
 * Handler for POST /v1/variation-envelopes
 * Operation: setVariationEnvelope
 */
export async function setVariationEnvelope(
  request: FastifyRequest,
  reply: FastifyReply,
  deps: CompositionDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: SetVariationEnvelopeInput = {
    ...(request.body ?? {}),
  };
  const result = await deps.useCases.variationEnvelopes.update.execute(input);
  return reply.code(201).send(result);
}