# platform/api-server/src/domains/experience/handlers/cohort.handlers.ts
/**
 * Handlers for Cohort
 *
 * Generated by Handler Generator v2.0.0
 * Generator Version: 2.0.0
 * Domain: experience
 *
 * ⚠️  DO NOT EDIT THIS FILE MANUALLY
 * This file is auto-generated. Any manual changes will be overwritten.
 */

import type { FastifyRequest, FastifyReply } from "fastify";
import type { ExperienceDomainModule } from "../dependencies/experience-ddd.dependencies.js";
import type { DefineCohortInput, ListCohortDefinitionsInput } from "@loadline/services/experience";
// Use cases are available via deps.useCases.<repoKey>.<verb>
// Execution context is automatically set up by route wrapper (wrapHandlerInExecutionContext)

/**
 * Handler for GET /v1/experience/cohorts
 * Operation: listCohortDefinitions
 */
export async function listCohortDefinitions(
  request: FastifyRequest<{ Params: Record<string, string>; Querystring: { cursor?: string | string[], limit?: string | string[], programmeId?: string | string[] }; Body: unknown }>,
  reply: FastifyReply,
  deps: ExperienceDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: ListCohortDefinitionsInput = {
    ...(request.query ?? {}),
  };
  const result = await deps.useCases.cohorts.list.execute(input);
  return reply.code(200).send(result);
}


/**
 * Handler for POST /v1/experience/cohorts
 * Operation: defineCohort
 */
export async function defineCohort(
  request: FastifyRequest,
  reply: FastifyReply,
  deps: ExperienceDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: DefineCohortInput = {
    ...(request.body ?? {}),
  };
  const result = await deps.useCases.cohorts.get.execute(input);
  return reply.code(201).send(result);
}