# platform/api-server/src/domains/permitting/handlers/factor-permit.handlers.ts
/**
 * Handlers for FactorPermit
 *
 * Generated by Handler Generator v2.0.0
 * Generator Version: 2.0.0
 * Domain: permitting
 *
 * ⚠️  DO NOT EDIT THIS FILE MANUALLY
 * This file is auto-generated. Any manual changes will be overwritten.
 */

import type { FastifyRequest, FastifyReply } from "fastify";
import type { PermittingDomainModule } from "../dependencies/permitting-ddd.dependencies.js";
import type { ListFactorPermitsInput, RecordFactorPermitInput } from "@loadline/services/permitting";
// Use cases are available via deps.useCases.<repoKey>.<verb>
// Execution context is automatically set up by route wrapper (wrapHandlerInExecutionContext)

/**
 * Handler for GET /v1/permitting/factor-permits
 * Operation: listFactorPermits
 */
export async function listFactorPermits(
  request: FastifyRequest<{ Params: Record<string, string>; Querystring: { cursor?: string | string[], limit?: string | string[], jurisdiction?: string | string[], position?: string | string[], dataCategory?: string | string[] }; Body: unknown }>,
  reply: FastifyReply,
  deps: PermittingDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: ListFactorPermitsInput = {
    ...(request.query ?? {}),
  };
  const result = await deps.useCases.factorPermits.list.execute(input);
  return reply.code(200).send(result);
}


/**
 * Handler for POST /v1/permitting/factor-permits
 * Operation: recordFactorPermit
 */
export async function recordFactorPermit(
  request: FastifyRequest,
  reply: FastifyReply,
  deps: PermittingDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: RecordFactorPermitInput = {
    ...(request.body ?? {}),
  };
  const result = await deps.useCases.factorPermits.get.execute(input);
  return reply.code(201).send(result);
}